+++ src/app/app.module.ts
@@ -13,6 +13,7 @@ import { SharedModule } from './shared/shared.module';
 import { AppRoutingModule } from './app-routing.module';
 import { PathNotFoundComponent } from './core/components/path-not-found/path-not-found.component';
 import { OrdersModule } from './orders/orders.module';
+import { httpInterceptorProviders } from './core/interceptors/index';
 
 @NgModule({
   declarations: [
@@ -30,7 +31,8 @@ import { OrdersModule } from './orders/orders.module';
     HttpClientModule,
     AppRoutingModule,
   ],
-  bootstrap: [AppComponent]
+  bootstrap: [ AppComponent ],
+  providers: [ httpInterceptorProviders ]
 })
 export class AppModule {
   constructor(router: Router) {