+++ src/app/core/models/app-settings.ts
@@ -1,0 +1,7 @@
+export class AppSettingsModel {
+  name: string;
+  company: string;
+  version: string;
+  date: Date;
+  author: string;
+}